Make version uint64

+++ pkg/crypto/vrf.go
@@ -9,7 +9,7 @@ import (
 
 type VrfInput struct {
 	Label   []byte // max length is 2^8-1, i.e., length can be stored in one byte
-	Version uint32
+	Version uint64
 }
 
 /*@
@@ -25,7 +25,7 @@ func encode(input VrfInput) (res []byte) {
 	buf := bytes.NewBuffer([]byte{})
 	buf.WriteByte(utils.Uint8(len(input.Label)))
 	buf.Write(input.Label)
-	buf.Write(utils.Uint32(input.Version))
+	buf.Write(utils.Uint64(input.Version))
 	return buf.Bytes()
 }
